% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellClassifier.R
\name{classify.cells}
\alias{classify.cells}
\title{Classify New Embedded Cells}
\usage{
classify.cells(data, classes, k = 7, seed = 18051982)
}
\arguments{
\item{data}{list; GFICF object}

\item{classes}{chareachters; Classes of aready exsiting cells in the order of thay are in colnames(data$gficf).}

\item{k}{integer; Number of K-nn to use for classification. Odd number less than 30 are prefered.}

\item{seed}{integer; Initial seed to use.}
}
\value{
A dataframe containing cell id and predicted classes.
}
\description{
Classify new embedded cells using GF-ICF transformation and K-nn algorithm.
Existing cells are used as training set.
}
