% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustCells.R
\name{clustcells}
\alias{clustcells}
\title{RphenoGraph clustering like}
\usage{
clustcells(data, from.embedded = F, k = 15,
  dist.method = "manhattan", nt = 2, community.algo = "louvian",
  store.graph = F, seed = 180582, verbose = TRUE)
}
\arguments{
\item{data}{list; Input data (gficf object)}

\item{from.embedded}{logical; Use embeddedd (UMAP or tSNA) space for clustering cells. Best results are usually obtained not using the embedded space.}

\item{k}{integer; number of nearest neighbours (default:15)}

\item{dist.method}{character; Dist to use for K-nn. Type of distance metric to use to find nearest neighbors. One of:
\itemize{
  \item \code{"euclidean"} (the default)
  \item \code{"cosine"}
  \item \code{"manhattan"}
  \item \code{"hamming"} (very slow)
}}

\item{nt}{integer; Number of cpus to use for k-nn search}

\item{community.algo}{characthers; Community algorithm to use for clustering. Supported are:
\itemize{
  \item \code{"louvian"} (the default)
  \item \code{"walktrap"}
  \item \code{"fastgreedy"}
}}

\item{store.graph}{logical; Store produced phenograph in the gficf object}

\item{seed}{integer; Seed to use for replication.}

\item{verbose}{logical; Increase verbosity.}
}
\value{
the updated gficf object
}
\description{
R implementation of the PhenoGraph algorithm
}
\details{
A custom R implementation of the PhenoGraph(http://www.cell.com/cell/abstract/S0092-8674(15)00637-6) algorithm, 
which is a clustering method designed for high-dimensional single-cell data analysis. It works by creating a graph ("network") representing 
phenotypic similarities between cells by calclating the Jaccard coefficient between nearest-neighbor sets, and then identifying communities 
using the well known Louvain method(https://sites.google.com/site/findcommunities/) in this graph.

That version used PCA or LSA reduced meta-cells and multithreading annoy version for K-nn search (from uwot package).
}
