% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensinalityReduction.R
\name{embedNewCells}
\alias{embedNewCells}
\title{Number of features to use}
\usage{
embedNewCells(data, x, nt = 2, seed = 18051982, ...)
}
\arguments{
\item{data}{list; GFICF object}

\item{x}{Matrix; UMI counts matrix of cells to embedd.}

\item{nt}{integer; Number of thread to use (default 2).}

\item{seed}{integer; Initial seed to use.}
}
\description{
Compute the number of dimension to use for either PCA or LSA.
}
