% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ensToSymbol}
\alias{ensToSymbol}
\title{Convert Ensamble IDs to Official Gene Symbols}
\usage{
ensToSymbol(df, col, organism, verbose = T)
}
\arguments{
\item{df}{data frame; Data frame containing the IDs to convert.}

\item{col}{characters; Name of column containing the ensamble ids.}

\item{organism}{characters; Organism of origin (i.e. human or mouse).}
}
\value{
The updated data frame with a new column called symb.
}
\description{
It uses biomart. If more the one gene is associated to the enamble, the first one retrived from
Biomart is used.
}
