% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gficf.R
\name{gficf}
\alias{gficf}
\title{Gene Frequency - Inverse Document Frequency (GFICF)}
\usage{
gficf(M, cell_proportion_max = 1, cell_proportion_min = 0.05,
  storeRaw = FALSE, normalize = FALSE)
}
\arguments{
\item{M}{Matrix; UMI cell count matrix}

\item{cell_proportion_max}{integer; Remove genes present in more then to the specifided proportion (0,1). Default 1.}

\item{cell_proportion_min}{integer; Remove genes present in less then or equal to the specifided proportion (0,1). Default is 0.05 (i.e. 5 percent).}

\item{storeRaw}{logical; Store RAW UMI counts.}

\item{normalize}{logical; Rescale UMI counts before applay GFICF. Recaling is done using EdgeR normalization.}
}
\value{
The updated gficf object.
}
\description{
R implementation of the GFICF (http://www.frontiers.path)
Thanks to 3’-end scRNA-seq approaches, we can now have an accurate estimation of gene expression without having to account for gene length,
thus the number of transcripts (i.e. UMI) associated to each gene, strictly reflects the frequency of a gene in a cell, exactly like a word in a document.
GFICF (Gene Frequency - Inverce Cell Frequency) is analugous of TF-IDF scoring method as defined for tex dada. With GFICF we consider a cell to be analogous
to a document, genes analogous to words and gene counts to be analogous of the word’s occurrence in a document.
}
