% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gficf.R
\name{loadGFICF}
\alias{loadGFICF}
\title{Restore GFICF object.}
\usage{
loadGFICF(file)
}
\arguments{
\item{file}{name of the file where the object is stored.}
}
\description{
Function to read a GFICF object from a file saved with \code{\link{saveGFICF}}.
}
\examples{
gficf_file <- tempfile("gficf_data")

# restore
data2 <- loadGFICF(file = gficf_file)

}
