% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCells}
\alias{plotCells}
\title{Plot cells in the ebedded space}
\usage{
plotCells(data, colorBy = NULL, pointSize = 0.5)
}
\arguments{
\item{data}{list; GFICF object}

\item{colorBy}{characters; Color cells according to a column contained in data$embedded data frame. Default is NULL.}

\item{pointSize}{integer; Size of the points in the plot. Default is 0.5.}
}
\value{
The updated gficf object.
}
\description{
Plot cells in the bidimensional space and color it according to a specific parameter.
}
