% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotGenes}
\alias{plotGenes}
\title{Plot gene expression across cells}
\usage{
plotGenes(data, genes, log2Expr = T, x = NULL, rescale = F)
}
\arguments{
\item{data}{list; GFICF object}

\item{genes}{characters; Id of genes to plot. It must correspond to the IDs on the rows of raw count matrix.}

\item{log2Expr}{boolean; Relative expression of a gene is computed on rescaled in log2 expression (default TRUE).}

\item{x}{Matrix; Custom normalized raw counts. If present will be used instead of the ones normalized by gficf. Default is NULL.}

\item{rescale}{boolean; Rescale expression between 0 and 1. Default is false.}
}
\value{
A list of plots.
}
\description{
Plot the expression of a group of genes across cells.
}
