% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotGeneViolin}
\alias{plotGeneViolin}
\title{Plot the expression of a gene across group of cells.}
\usage{
plotGeneViolin(data, gene, ncol = 3, x = NULL)
}
\arguments{
\item{data}{list; GFICF object}

\item{gene}{characters; Id of genes to plot. It must correspond to the IDs on the rows of raw count matrix.}

\item{ncol}{numeric; Number of columns of the final plot (defaul is 3).}

\item{x}{Matrix; Custom normalized raw counts. If present will be used instead of the ones normalized by gficf. Default is NULL.}
}
\value{
A list of plots.
}
\description{
Plot the expression of a gene across group of cells with violion plot.
}
