% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotGSVA}
\alias{plotGSVA}
\title{Plot GSEA results}
\usage{
plotGSVA(data, fdr = 0.05, clusterRowCol = T, logFCth = 0)
}
\arguments{
\item{data}{list; GFICF object}

\item{fdr}{number; FDR threshold to select significant pathways to plot.}

\item{clusterRowCol}{boolean; if TRUE row and col of the plot are clustered.}

\item{logFCth}{number; LogFC threshold to select pathways to plot.}
}
\value{
plot from ggplot2 package.
}
\description{
Circle plot for gene set enrichement analysis results.
}
