% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPathway}
\alias{plotPathway}
\title{Plot GSEA results}
\usage{
plotPathway(data, pathwayName, fdr = 0.05)
}
\arguments{
\item{data}{list; GFICF object}

\item{pathwayName}{characters; Name of the pathway to plot.}

\item{fdr}{number; FDR threshold to select significant pathways to plot.}
}
\value{
plot from ggplot2 package.
}
\description{
Circle plot for gene set enrichement analysis results.
}
