% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensinalityReduction.R
\name{runLSA}
\alias{runLSA}
\title{Latent Semantic Analysis (LSA)}
\usage{
runLSA(data, dim = NULL, rescale = F, centre = F, randomized = T)
}
\arguments{
\item{data}{list; GFICF object}

\item{dim}{integer; Number of dimension which to reduce the dataset.}

\item{rescale}{logical; Rescale gficf scores before applying reduction (deprecated).}

\item{centre}{logical; Centre gficf scores before applying reduction (increase separation).}

\item{randomized}{logical; Use randomized (faster) version for matrix decomposition (default is TRUE).}
}
\value{
The updated gficf object.
}
\description{
Reduce dimensionality of the single cell dataset using Latent Semantic Analysis (LSA)
}
