% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoTime.R
\name{runPseudoTime}
\alias{runPseudoTime}
\title{Pseudotime with slingshot}
\usage{
runPseudoTime(
  data,
  hgv = T,
  parallel = F,
  verbose = T,
  start.cl = NULL,
  end.cl = NULL,
  ...
)
}
\arguments{
\item{data}{list; GFICF object}

\item{hgv}{boolean; use only top High Variable Genes (FDR<.25)}

\item{parallel}{boolean; Parralle execution}

\item{verbose}{boolean; Increase verbosity}

\item{start.cl}{integer; (optional) character, indicates the cluster(s) from which lineages will be drawn.}

\item{end.cl}{integer; (optional) character, indicates the cluster(s) which will be forced leaf nodes in their trees.}

\item{...}{Additional arguments to pass to fitGAM of tradeSeq.}
}
\value{
The updated gficf object.
}
\description{
The function is just a wrapper for slingshot and tradeSeq packages. 
Can be used to assign a pseudotime to each cell. Identified clusters in the UMAP space will be used.
}
