% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gficf.R
\name{saveGFICF}
\alias{saveGFICF}
\title{Save GFICF object.}
\usage{
saveGFICF(data, file = "~/Documents/gficf.test")
}
\arguments{
\item{data}{a GFICF object create by \code{\link{gficf}}.}

\item{file}{name of the file where the model is to be saved.}
}
\description{
Function to write a GFICF object to a file, preserving UMAP model.
}
\examples{
# save
gficf_file <- tempfile("gficf_test")
saveGFICF(data, file = gficf_file)

# restore
data2 <- loadGFICF(file = gficf_file)

}
