% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensinalityReduction.R
\name{embedNewCells}
\alias{embedNewCells}
\title{Embed new cells in an existing space}
\usage{
embedNewCells(data, x, nt = 2, seed = 18051982, verbose = TRUE, ...)
}
\arguments{
\item{data}{list; GFICF object}

\item{x}{Matrix; UMI counts matrix of cells to embedd.}

\item{nt}{integer; Number of thread to use (default 2).}

\item{seed}{integer; Initial seed to use.}

\item{verbose}{boolean; Icrease verbosity.}

\item{...}{Additional arguments to pass to Rtsne or umap_transform call.}
}
\value{
The updated gficf object.
}
\description{
This function embed new cells in an already existing space. For now it supports only UMAP and t-UMAP. Briefly new cells are first normalized with GF-ICF method but using as ICF weigth estimated on the existing cells and than projected in the existing PCA/LSA space before to be embedded in the already existing UMAP space via umap_transform function.
}
