% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deGenes.R
\name{findClusterMarkers}
\alias{findClusterMarkers}
\title{Find Marker Genes from cell clusters.}
\usage{
findClusterMarkers(data, nt = 2, hvg = T, verbose = T)
}
\arguments{
\item{data}{list; GFICF object}

\item{nt}{integer; Number of thread to use (default 2).}

\item{hvg}{boolean; Use only High Variable Genes (default is TRUE).}

\item{verbose}{boolean; Icrease verbosity.}

\item{x}{Matrix; UMI counts matrix of cells to embedd.}
}
\value{
The updated gficf object.
}
\description{
Try to identify marker genes across clusters performing Mann-Whitney U test.
DE genes are identified the expressio in each cluster versus the all the other.
}
