% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deGenes.R
\name{findVarGenes}
\alias{findVarGenes}
\title{Find high variable genes following the approach
 proposed by Chen et al. in BMC Genomics (2016)
 Code adapted from https://github.com/hillas/scVEGs}
\usage{
findVarGenes(data, fitMethod = "locfit", verbose = T)
}
\description{
Find high variable genes following the approach
 proposed by Chen et al. in BMC Genomics (2016)
 Code adapted from https://github.com/hillas/scVEGs
}
