% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnalisys.R
\name{runGSEA}
\alias{runGSEA}
\title{Gene Set Enrichement Analysi on GF-ICF}
\usage{
runGSEA(data, gmt.file, nsim = 1000, convertToEns = T,
  convertHu2Mm = F, nt = 2, minSize = 15, maxSize = Inf,
  verbose = TRUE, seed = 180582)
}
\arguments{
\item{data}{list; GFICF object}

\item{gmt.file}{characters; Path to gmt file from MSigDB}

\item{nsim}{integer; number of simulation used to compute ES significance.}

\item{convertToEns}{boolean: Convert gene sets from gene symbols to Ensable id.}

\item{convertHu2Mm}{boolean: Convert gene sets from human symbols to Mouse Ensable id.}

\item{nt}{numeric; Number of cpu to use for the GSEA}

\item{minSize}{numeric; Minimal size of a gene set to test (default 15). All pathways below the threshold are excluded.}

\item{maxSize}{numeric; Maximal size of a gene set to test (default Inf). All pathways above the threshold are excluded.}

\item{verbose}{boolean; Show the progress bar.}

\item{seed;}{Seed to use for random number generation.}
}
\value{
The updated gficf object.
}
\description{
Compute GSEA for each cluster across a set of input pathways.
}
